requirements_backend.txt
# ============================================
# TRADING BOT PRO - BACKEND DEPENDENCIES
# ============================================

# Core Web Framework
flask==3.0.0
flask-socketio==5.3.5
flask-cors==4.0.0

# Security & Authentication
bcrypt==4.1.2
PyJWT==2.8.0
cryptography==41.0.7

# Database
# SQLite is built-in to Python, no external package needed

# Email
# Built-in smtplib and email modules (no external package needed)

# Trading & APIs
pybit==5.7.0
python-telegram-bot==20.7
telethon==1.33.1

# HTTP & Requests
requests==2.31.0
urllib3==2.1.0

# WebSocket Support
python-engineio==4.8.0
python-socketio==5.10.0

# Environment Variables
python-dotenv==1.0.0

# Data Validation
email-validator==2.1.0

# Rate Limiting (optional)
Flask-Limiter==3.5.0

# Utilities
python-dateutil==2.8.2

# Development Tools (optional)
# pytest==7.4.3
# pytest-cov==4.1.0
